# api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Folo Hooks API",
    "version": "1.0.0",
    "description": "代理转发接口文档"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "本地开发服务器"
    },
    {
      "url": "https://atoolio.com",
      "description": "测试服务器"
    }
  ],
  "paths": {
    "/folo/search": {
      "get": {
        "summary": "搜索接口",
        "description": "代理转发到本地搜索接口",
        "parameters": [
          {
            "name": "title",
            "in": "query",
            "description": "按标题模糊搜索",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "author",
            "in": "query",
            "description": "按作者模糊搜索",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "url",
            "in": "query",
            "description": "按URL模糊搜索",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "页码，默认为1",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "每页条数，默认10条，最大100条",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            }
          },
          {
            "name": "startTime",
            "in": "query",
            "description": "开始时间，格式：YYYY-MM-DD HH:mm:ss",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{2}-\\d{2} \\d{2}:\\d{2}:\\d{2}$",
              "example": "2024-01-01 00:00:00"
            }
          },
          {
            "name": "endTime",
            "in": "query",
            "description": "结束时间，格式：YYYY-MM-DD HH:mm:ss",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{2}-\\d{2} \\d{2}:\\d{2}:\\d{2}$",
              "example": "2024-12-31 23:59:59"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功返回搜索结果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "total": {
                      "type": "integer",
                      "description": "总记录数"
                    },
                    "page": {
                      "type": "integer",
                      "description": "当前页码"
                    },
                    "limit": {
                      "type": "integer",
                      "description": "每页条数"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/articles/genguonei": {
      "get": {
        "summary": "跟国内接口",
        "description": "获取最近12小时的国内新闻数据",
        "tags": ["Articles"],
        "parameters": [
          {
            "name": "title",
            "in": "query",
            "description": "按标题模糊搜索",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "author",
            "in": "query",
            "description": "按作者模糊搜索",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "url",
            "in": "query",
            "description": "按URL模糊搜索",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "页码，默认为1",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
             "name": "limit",
             "in": "query",
             "description": "每页条数，默认100条",
             "required": false,
             "schema": {
               "type": "integer",
               "minimum": 1,
               "maximum": 1000,
               "default": 100
             }
           },
           {
             "name": "endTime",
             "in": "query",
             "description": "结束时间，格式：YYYY-MM-DD HH:mm:ss（可选，默认为当前时间）",
             "required": false,
             "schema": {
               "type": "string",
               "pattern": "^\\d{4}-\\d{2}-\\d{2} \\d{2}:\\d{2}:\\d{2}$",
               "example": "2024-12-31 23:59:59"
             }
           }
         ],
        "responses": {
          "200": {
            "description": "成功返回国内新闻数据",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "请求是否成功"
                    },
                    "message": {
                      "type": "string",
                      "description": "响应消息"
                    },
                    "timestamp": {
                      "type": "string",
                      "description": "响应时间戳",
                      "example": "2023-12-21 15:30:00"
                    },
                    "searchData": {
                      "type": "object",
                      "description": "搜索返回的数据"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "调用search接口失败"
                    },
                    "error": {
                      "type": "string",
                      "description": "错误详情"
                    },
                    "timestamp": {
                      "type": "string",
                      "description": "错误时间戳"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/articles/genguoji": {
      "get": {
        "summary": "跟国际接口",
        "description": "获取最近12小时的国际新闻数据",
        "tags": ["Articles"],
        "parameters": [
          {
            "name": "title",
            "in": "query",
            "description": "按标题模糊搜索",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "author",
            "in": "query",
            "description": "按作者模糊搜索",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "url",
            "in": "query",
            "description": "按URL模糊搜索",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "页码，默认为1",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "每页条数，默认100条",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 100
            }
          },
          {
            "name": "endTime",
            "in": "query",
            "description": "结束时间，格式：YYYY-MM-DD HH:mm:ss（可选，默认为当前时间）",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{2}-\\d{2} \\d{2}:\\d{2}:\\d{2}$",
              "example": "2024-12-31 23:59:59"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功返回国际新闻数据",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "请求是否成功"
                    },
                    "message": {
                      "type": "string",
                      "description": "响应消息"
                    },
                    "timestamp": {
                      "type": "string",
                      "description": "响应时间戳",
                      "example": "2023-12-21 15:30:00"
                    },
                    "searchData": {
                      "type": "object",
                      "description": "搜索返回的数据"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "调用search接口失败"
                    },
                    "error": {
                      "type": "string",
                      "description": "错误详情"
                    },
                    "timestamp": {
                      "type": "string",
                      "description": "错误时间戳"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
   }
 }